backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
numpy>=1.21.0